@@ -3,8 +3,6 @@ import NavBar from '@/components/NavBar'
 import EditorBox from '@/components/editor/EditorBox'
 import GameBox from '@/components/game/GameBox'
 import ConsoleBox from '@/components/console/ConsoleBox'
-import AppParametersModal from '@/components/game/AppParametersModal'
-import { mapState } from 'vuex'
 /**
  * Base Vue component
  */
@@ -13,13 +11,7 @@ export default {
     NavBar,
     EditorBox,
     GameBox,
-    ConsoleBox,
-    AppParametersModal
-  },
-  computed: {
-    ...mapState({
-      showAppParametersModal: state => state.showAppParametersModal
-    })
+    ConsoleBox
   }
 }
 </script>
@@ -30,6 +22,5 @@ export default {
     <EditorBox />
     <GameBox />
     <ConsoleBox />
-    <AppParametersModal v-if="showAppParametersModal" />
   </div>
 </template>
