@@ -9,8 +9,8 @@ export default {
   },
   mounted () {
     // When the box for the game is mounted, we can load the default game
-    this.loadGame()
     this.$store.dispatch('console/info', 'Game box mounted, ready to load game !')
+    this.loadGame()
   },
   methods: {
     ...mapActions({
