@@ -1,3 +1,5 @@
+import { GameManifestValidator } from './GameManifestValidator'
+
 /**
  * The GameManager is the orchestrator of the game
  * He is responsible of :
@@ -23,15 +25,26 @@ export default class GameManager {
   async loadGame () {
     // Loads the manifest of the game (from the external source)
     await this.provider.loadManifest()
+    // Check the manifest validity
+    let validator = new GameManifestValidator(this.provider.manifest)
+    // Validate format
+    validator.checkFormat()
+    // Check if specified languages are compatible
+    validator.checkLanguages()
     // Loads the available interpreters
     await this.provider.loadInterpreters()
+    // Loads game libraries
+    await this.provider.loadLibraries()
     // Loads game classes
     await this.provider.loadGameClass()
   }
   /**
-   * Loads the game
+   * Responsible to start the game
+   * This method instanciate the game base class declared in the game module
+   *
+   * Change the game api, this method must be asynchronus
    */
-  async startGame () {
+  startGame () {
     // Create the base game class instance
     window.game = new this.provider.gameModule.Game(
       document.getElementById('game-box'),
