@@ -4,7 +4,6 @@
  * @author Bastien Nicoud
  */
 export default {
-  showAppParametersModal: false,
   /**
    * Stores all editor related datas
    * @type {Object}
@@ -20,6 +19,11 @@ export default {
      * @type {String}
      */
     language: localStorage.getItem('editor-language') || 'ruby',
+    /**
+     * Default font size
+     * @type {Number}
+     */
+    fontSize: localStorage.getItem('editor-font-size') || 14,
     // Mapping between the language and the content of the editor
     languagesContent: {
       javascript: localStorage.getItem('javascript') || `let tutu = 'Bienvenue sur LiveCoding'\ndebugger.log(tutu)\n`,
@@ -29,14 +33,15 @@ export default {
     }
   },
   processor: {
-    url: localStorage.getItem('processor-url') || 'ws://localhost:8787/processor'
+    url: localStorage.getItem('processor-url') || 'ws://localhost:12800/'
   },
   /**
    * Information form the current loaded game
    */
   game: {
-    provider: localStorage.getItem('game-provider') || 'url',
-    url: localStorage.getItem('game-url') || 'http://localhost:3333',
+    loading: false,
+    provider: localStorage.getItem('game-provider') || 'github',
+    url: localStorage.getItem('game-url') || 'CPNV-ES/LiveCoding-Test-Game@v1.1.8',
     loaded: false,
     manager: null
   }
